requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
praw==7.7.1
prawcore==2.4.0
psaw==0.1.0
python-dotenv==1.0.0
requests==2.31.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.6.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
praw==7.7.1
prawcore==2.4.0
psaw==0.1.0
python-dotenv==1.0.0
rake-nltk==1.0.6
regex==2023.10.3
requests==2.31.0
tqdm==4.66.1
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.6.4
